Function to get zipcodes within x miles of user.
