+++ .gitignore
@@ -2,3 +2,4 @@ node_modules
 .DS_STORE
 public/bundle.js*
 resources/courseLists/*
+.env*
